+++ src/ApiClient/ProviderCsvApiClient.php
@@ -10,9 +10,7 @@
 
 namespace PayBreak\Sdk\ApiClient;
 
-use Psr\Http\Message\ResponseInterface;
-use WNowicki\Generic\ApiClient\ErrorResponseException;
-use WNowicki\Generic\ApiClient\WrongResponseException;
+use PayBreak\ApiClient\CsvApiClient;
 use Psr\Log\LoggerInterface;
 
 /**
@@ -21,7 +19,7 @@ use Psr\Log\LoggerInterface;
  * @author EA
  * @package PayBreak\Sdk\ApiClient
  */
-class ProviderCsvApiClient extends ProviderApiClient
+class ProviderCsvApiClient extends CsvApiClient
 {
     /**
      * @author WN
@@ -32,47 +30,12 @@ class ProviderCsvApiClient extends ProviderApiClient
      */
     public static function make($baseUrl, $token = '', LoggerInterface $logger = null)
     {
-        $ar = [];
-        $ar['base_uri'] = $baseUrl;
+        $headers = [];
 
         if ($token != '') {
-
-            $ar['headers'] = ['Authorization' => 'ApiToken token="' . $token . '"'];
+            $headers['Authorization'] = "ApiToken token=\"" . $token . "\"";
         }
 
-        return new self($ar, $logger);
-    }
-
-    /**
-     * @author EA
-     * @param ResponseInterface $response
-     * @return array
-     * @throws WrongResponseException
-     */
-    protected function processResponse(ResponseInterface $response)
-    {
-        if ($response->getStatusCode() == 204) {
-            return null;
-        }
-
-        if (strpos($response->getHeaderLine('Content-Type'), 'text/csv') !== false) {
-            return ['csv' => $response->getBody()->getContents()] ;
-        }
-
-        throw new WrongResponseException('Response body was malformed csv', $response->getStatusCode());
-    }
-
-    /**
-     * @author EA
-     * @param string $uri
-     * @param array $body
-     * @param array $query
-     * @return array
-     * @throws ErrorResponseException
-     * @throws \Exception
-     */
-    public function get($uri, array $body = [], array $query = [])
-    {
-        return parent::get($uri . '.csv', $body, $query);
+        return new self(['base_uri' => $baseUrl], $logger, $headers);
     }
 }