+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 4.14.1
+2017-06-15
+
+### Features
+- Added `Liability` concept on `Application` and `Installation` entity (147092245)
+
+# 4.14.0
+2017-05-03
+
+### Features
+- Use `PayBreak\ApiClient` instead of `Wnowicki\Generic\ApiClient` (143560655)
+
 # 4.13.0
 2017-04-03